Schedd should decline executable spooling if file already exists. #6454
When copy_to_spool is set to true in the submit file, condor_submit
attempts to send the executable to the schedd once for every job. Since
the spooled executable is shared across all jobs within a cluster, this
can result in many needless transfers.
If submit uses the SendSpoolFileIfNeeded() RPC, the schedd will now
decline dpulicate transfers.