Maintain symmetry of version information between cedar peers. #6436
When using a cached security session that lacks peer version information
for a new connection, the server should not copy the client version info
from the resume-session ad into the sock object. The client's sock
object won't have peer version info, and we want to maintain symmetry in
the presence or absence of that data between client and server. We rely
on that to make backwards-compatibile changes to the protocol of some
cedar commands.