Remove old version check for renaming of stdout/err. #6436
The syscall_sock doesn't have version info for the shadow's version
check when SEC_ENABLE_MATCH_PASSWORD_AUTHENTICATION=True (the default).
Without the check, if the starter or shadow is older than 7.7.2, the
job's stdout/err may be written to the wrong location if the filenames
in the job ad contain any path information.
Removing this backwards-compatibility change seems unlikely to harm
anyone.