Fix interference between systemd and master restart code. #6476
When the master exec()s itself, information from systemd is lost
(primarily environment variables). As a result, the new master image
doesn't know that it needs to send watchdog keepalive messages or that
it was given a previously-bound command port. In this case, the msater
now exits, expecting systemd to restart it.
Also, when not running under systemd, the master no longer wakes up
every second to try (and fail) to send a keepalive to systemd.