Add blahp options for project name and runtime limit. #6451
These are two common job attributes in the batch systems supported by
the blahp. These attributes are specific to to the batch type under the
grid universe. BatchProject is a string that's used with the project or
allocation option of the batch system's submission interface.
BatchRuntime is an integer that limits the job's wallclock execution
time, expressed in seconds.