Remove Google coredumper. #6482
The Google coredumper doesn't write usable core files on rhel7. It
hasn't had any updates in many years, so we're unlikely to see a fix.
Now that we can make the kernel write core files for daemons that change
uid, we can use that instead. The small drawback is that we cannot
control the names of the core files (config param CORE_FILE_NAME).