Allow ignoring of IPv4 reults from getaddrinfo(). #4881
IPv4 results are no longer returned by ipv6_getaddrinfo() when
ENABLE_IPV4=False.
Filtering of IPv4 and IPv6 results is now done via the hint paremeter in
getaddrinfo(), instead of in the iterator of results.
Before this patch, clients with ENABLE_IPV4=False would attempt to use
IPv4 to contact a server whose location was given as a hostname (e.g.
the collector).