Add config knob to constrain which jobs the negotiator will consider. #6250
The new knob, NEGOTIATOR_JOB_CONSTRAINT, is an expression evaluated
against the job. If False, then the negotiator doesn't want to consider
the job for matchmaking. The knob is read by the negotiator and then
sent to the schedd during negotiation. The schedd will not send jobs to
the negotiator which don't match the constraint.
Older schedds will ignore this constraint, and at present, the
negotiator doesn't apply it to request ads that it receives.