Re-optimize pslot ad after mutating it into a dslot ad. #6593
Optimization of a pslot ad will copy the Cpus, Memory, and Disk resource
values into the Requirements expression. If we modified these values as
part of making the pslot look like a dslot, we then need to redo the
optimization.
Otherwise, the optimized Requirements expression will allow jobs that
fit in the original pslot but not in the dslot. This leads to
relinquished claims when the startd refuses to run the jobs.