Remove CEDAR's send-a-NULL-as-a-string feature. #6589
Some of the methods in CEDAR to send a string support sending a NULL
string pointer, which the receiver will get instead of an allocated
buffer. Many receivers of strings from CEDAR are not expecting a
possible NULL pointer, and will crash if they get one.
Remove this feature from the standard CEDAR communication methods. For
the two places that rely on this behavior (remote syscalls and
config_val), add alternate methods that others should avoid.
Receivers will still recognize the special NULL pattern, but will now
translate it into a zero-length string.