Job can leave queue without final ad updates by schedd. #6626
Don't do periodic job policy evaluation for HELD or COMPLETED jobs if
they have a shadow_rec.
If a shadow asks to be recycled but there's no new job for it, the old
job's status can be changed to HELD or COMPLETED but other updates to
the job ad (like RemoteWallClockTime) don't happen until the reaper for
the shadow fires. A policy evaluation in between can cause bad things to
happen.