condor_transfer_data writes user log to wrong directory. #6695
When condor_transfer_data fetches a job's files, the schedd sends just
the basename for the user log file. This results in the client always
writing the file in the job's IWD, even if the original job ad said to
write the file in another directory.
Add an automatic file transfer remap in the client so that the file ends
up in the right place.