Fix use-after-delete bug in ResMgr destructor. #6345
m_attrs is deleted, then ~Claim() calls Resource::dprintf_va(), which
calls m_attrs->is_smp() to decide if it should add the slot id as a
prefix.
I think we can assume that all machines are SMPs now.
And move the delete of m_attrs to catch other uses of it during
destruction.