Handle full sinful strings in claim ids. #6810
We want to be able to put the startd's full sinful string in its claim
ids. But a sinful string can contain '#', which is also the field
separator in a claim id. Thus, we should look for '>' to find the end of
the sinful string in a claim id instead of '#'.