Add condor_ssh_to_job support for Singularity #6825
Allow condor_docker_enter to work for singularity.  It connects to a
unix domain socket the starter has created in the scratch dir, and
which the starter is listening on.  At connection, condor_docker enter
passes stdin, stdout and stderr to the starter, which as root execs
usr/bin/nsenter -a -t <jobPid> /usr/sbin/chroot root_dir with the
passed stdin, stdout and sterr dup'ed into the right places.