First pass at allowing the starter to forward its job elsewhere. #6843
Add RemoteProc to the starter, which allows it to send its job somewhere
else for execution. The actual movement of the job and status updates is
done by an external program called by the starter.
Neither the startd nor the shadow need to be aware that the job has been
diverted.