Only match against the "suffix" of the URL scheme.  GT #6858
This causes the URL matching for the file transfer plugins to
only look at the "suffix" of the URL scheme (everything after
the last `+` character).  So a URL of the form `foo+https://` would
match against a plugin advertising support for `https`.

By looking at a subset of the scheme, this allows us to introduce some
semantics in the scheme prefix later without having to register every
possible variation of the scheme prefix.