On recycle, shadow can log new job's events under old job's id. #6919
When the shadow is recycled, it deletes all of the data structures for
the previous job and creates new ones for the next job. The WriteUserLog
object doesn't get recreated, since it's a static member of BaseShadow.
If the next job doesn't require any logging, then no initialization
happens to the WriteUserLog object, and it will write events for the
next job to the old log files with the previous job's id.

The fix is to make the WriteUserLog object a non-static data member and
recreate it for each job. This requires an adjustment to the exception
logging code for the case where the BaseShadow object isn't present.