Fix evaluation functions in compat ClassAds layer. #6852
Allow the underlying ClassAds code to do type conversions between
boolean, integer, and real values when doing an expression evaluation
with a specific type in mind for the result.
This simplifies the compat code and makes it act more consistently with
the way type conversions occur within an expression evaluation.
For example, the old compat Lookup() calls wouldn't convert a result
from real to integer or from boolean to real. But these conversions do
occur within an expression.

Also fix a bug in EvalString() where evaluation is done on the wrong ad.

For now, I'm leaving in place the questionable code in the Eval*()
functions that tries to evaluate the given attribute in the target ad if
it's not in the local ad. But code that's doing these calls should know
which ad their attribute is expected in.