Add support for multiple keys. #6947
Add support on the server for advertising the keys and issuer available
during the authentication negotiation.  This enables the client to
determine which derived key to send to the server.

Adds support in both the server and client portions to fetch the
correct key for signing and verifying.

Add support in the client CLI to have the key specified as an option.