Add support for fetching session token from daemons. #6947
This adds a new daemon core command, DC_GET_SESSION_TOKEN, which
returns a signed token equivalent to the session established with the
daemon.  This token can subsequently be used by a client to authenticate
via the PASSWORD2 method and receive the same identity.

If the session was authenticated with PASSWORD2 itself and the token used
to establish the current session has a set expiration date, then the
resulting token will retain the same expiration date.