Speed up filtering of private ads.  #6956
Now that the collector always filters private attributes from ads
when sending to remote clients, we should make sure this is not
too expensive.

This tweak removes the cost of filtering private ads *in the case*
where we have no private attributes - this should always be true
for the large bulk (public startd ads) of the collector ads.

It also tweaks the private ad data structure to be a hash map instead
of a map.