Have the schedd create a capability for negotiation. #6956
This causes the schedd to create a capability, good for a remote
host to negotiate with it, and adds it into the submitter ads.

With this, the remote negotiator simply needs to look at the private
attributes of the submitter ad and use it to create a new security
session.  This essentially causes the schedd to trust any negotiator
that is trusted by its collector(s).