Add low-speed timeouts to the curl plugin. #6971
With this, curl will timeout if the transfer is going too slow
(default is <1KB/s for over 30 seconds).  The default is meant to
catch truly dead hosts, not provide a minimum performance.

However, the curl plugin will also parse the job and machine ad where
the default limits can be overridden.