Use the error code, not retry number.  #6971
The curl plugin appears to have a frustrating bug that, when it checks
the exit code for a failure that should be retried, it doesn't actually
check the exit code - rather, it checks the retry number!

Accordingly, without this fix, the curl plugin will never retry a failure,
regardless of the cause.