Make ClassAd lexer less greedy in reading unneeded characters. #7002
The lexer now fetches characters from the LexerSource only when it needs
them to resolve the current token. This reduces the numbr of places it
need to unread a character. This is visible to callers in that after
parsing a ClassAd, a FILE source won't have a character pushed back via
ungetc().