Keep CEDAR peers in sync after GSI authentication fails. #7024
Remove a workaround for a GSI bug that can leave the peers out of sync
when the server's subject name doesn't match what the client is
expecting. The GSI server code is left expecting data that the
client code doesn't send.
We tell GSI that the client should accept any subject for the
server (since we'll check it afterwards), so this case shouldn't arise.
The workaround puts the peers out of sync if GSI fails for any other
reason.