Have the server send a failure indicator for bad tokens. #7011
This refactors some of the server logic so we parse the token
before the last network message is passed.  This way, if the server
doesn't like the token itself (invalid format, expired, etc), it
will send a failure back to the client.

Previously, the failure happened too late and the client and server would
end up disagreeing on the network state (causing the client to eventually
timeout).