Allow jobs to use transfer plugins on evict. #7042
With this change, the job can set `OutputPluginsOnlyOnExit` to
`false` (defaults to `true`) to cause file transfer plugins to
be used whenever the job stages out.

The use case here is to allow self-checkpointing jobs to move
data to somewhere besides the schedd spool (as the checkpoints are
potentially quite large).  It is currently up to the user to make
sure the checkpoints are copied back into the working directory.

Note that this requires us to calculate the download remaps when we
determine that we want to enable URL mapping and spools.