Improve code to keep CEDAR peers in sync afer GSI failures. #7024
It turns out the workaround for a GSI bug that we recently removed is
necessary in some situations that we encounter. When the SSL handshake
fails in certain ways, gss_init_sec_context() accidentally drops the
last packet to be sent to the server by the client. In other failure
cases, the final data packet is sent.
We are restoring the old workaround, but limit the situations when it's
triggered. Now, it only occurs when the last packet exchanged is a
"large" packet received by the client. This correctly handles the
situations where we've seen this problem occur.