Do not fail if initd cannot be removed completely.
When run on NFS, this directory cannot be deleted because the
perl process itself holds an open file handle to a file inside
the directory, resulting in undeleteable shadow file handles
until the test suite fails.

It's not relevant to the test to have this fail hard when
cleanup fails, so just allow thigns to proceed.