Introduce the `FlockTo` attributes.
This attribute controls where the jobs are interesting in flocking;
there is a special value, `default`, that indicates the job is interested
in whatever targets are set by the sysadmin.

The schedd will now calculate both idle and running jobs based on the
value of this attribute - and whether the admin has enabled flocking by
default for all jobs.