Allow a session policy to be stored with security sessions.
The session policy can include data (such as the X509 auth attributes
or the remote pool the non-negotiated session was associated with)
that is stored as part of the generated session - but not sent to
the remote side.  This change allows the session data to be more easily
stored when generating non-negotiated sessions.