Override boolean coercion of `Value` to do something reasonable.  #7109
The boolean coercion of `Value`, as it is a boost enum, defaulted to
integer coercion.  Thus, `bool(classad.Value.Undefined)` returns `True` -
a horribly surprising result.

It turns out that it is not overly difficult to overload operators for
enums; it's just not well-advertised.  This does the overload in the case
of boolean coercion.  Others may be reasonable in the future.