Add special cases to restore backward compat. #7109
With this, once again `Undefined == Undefined` evaluates to True.
From a ClassAd perspective, this is the incorrect thing to do -- it
should become False since Undefined is coerced to False.

However, this broke a lot of unit tests and I'm concerned about real-world
backward compatibility impact.

That said, Error comparisons should still cause an exception; given the
expected rarety of these, I'm OK with the change.  Unit tests were updated
to reflect the new behavior.