Have token request utilize callbacks. #7093
With this, we utilize the new callback interface when sending
updates to the collector.  On failure, we have the ability to
trigger a token request (if the daemon had setup such a thing
beforehand).

On the successful installation of a token, the daemon can get
a callback where it can take appropriate action to start using
the token (typically, reset the collector update timer).