Make token request infrastructure 'identity-aware'.
The token request infrastructure had some of the plumbing necessary
to do user-specific token requests; this commit finishes up the
interfaces and implementation to make it all work.

With this, it is possible for a daemon to request a token for a
non-root user and install the token to the correct location.