Allow users to approve some token requests. #7159
This allows users to approve token requests for their own identity.
It is semantically equivalent to having the user generate their own
token (which is currently allowed) followed by copy/pasting it to
a remote machine.

With this work, token generation can be done completely without the
user ever "seeing" the token; if the user never touches the token
contents, then it reduces the likelihood they'll copy/paste it into
a Jupyter notebook and commit it to GitHub!