Implement Token / TokenRequest API in python. #7162
Provides a straightforward way to request and utilize tokens via
python.  The `TokenRequest` object takes its inspiration from a
`Future` object in python.

The `Token` object allows the user to write the contained token
to disk or utilize it as part of a new security session.  However,
there's no easy way to dump the contents to the console (or a
Jupyter notebook for that matter...).  The `Token` object is
pickle-able, meaning that it can be passed via processes in
multiprocessing-based environments.