condor_read() should recognize abnormal socket closure. #7164
condor_read() now returns -2 on abnormal socket closure (recv() fails
with ECONNRESET or similar). Before, it only did so on orderly socket
closure (recv() returns 0).
This allows ReliSock::readReady() to accurately indicate that the sock
is ready for reading (due to the connection being closed), which is
important for non-blocking operations.
Before this change, non-blocking readers could end up in a busy loop
checking the socket for a CEDAR message until the timeout is reached.