Startd can't update starter when EncryptExecuteDirectory=true #7169
Starter::executeDir() now always returns the value of parameter EXECUTE
configured for the given starter process. When an encrypted workspace is
requested, this will be different than the EXECUTE value for slot.
This fixes a bug where the startd fails to write a .update.ad file in
the job's execute directory.