Initial support for kubernetes under the blahp. #7202
Add blahp scripts for running jobs in a kubernetes cluster. Assumes the
kubectl command is in the path and can be run without parameters for
configuration or credentials.
Most pod parameters (including memory and image) are hard-coded and no
file transfer is done.
The hold and resume commands and most options to submit are not
supported yet.