Add WriteUserLog::initialize() based on a job ad. #5329
The code in this function is currently replicated sloppily in a half
dozen locations. That code can be consolidated here. We configure the
user, dagman, and global event logs as appropriate. This includes
checking the xml and dagman event mask attributes. Priv switching to
user and condor will be done for file I/O. Optionally, the caller can
request that init_user_ids() be called.

Also, move getPathToUserLog() to prevent adding more files to
libcondorapi.
And don't switch to condor priv when writing to the user or dagman logs.
And have initialize() that takes a single filename disable writing to
the global event log.