Build RTD with most mature bindings released on PyPI for given version #7263
Uses the fact that the earliest pre-release version string allowed on PyPI
is (or will be normalized to) ${VERSION}a0, and we impose a hopefully
reasonable maximum on post-releases to 999 (i.e. ${VERSION}.post999 in
PyPI-normalized-version-string language).

This commit also relaxes the regex that pulls the version string from
CMakeLists.txt in case anyone ever goes in and adds whitespace before
or spaces inside the set() call as is in done in random places in
other CMakeLists.txt files in our code.