More improvements to setuptools install of python bindings #6907
We no longer output python bindings so-files to a different directory,
as setuptools can build and install the python libraries' scaffolding
to the correct locations without the so-files existing in the source
tree. Note that the install target of the so-files is slightly
different now, either ${C_PYTHONARCHLIB}/{classad,htcondor}. As a bonus,
this means the MacOS path fixing can work pretty much as normal.

Also, now the cmake variable C_PYTHONARCHLIB points to something like
lib/pythonX.Y/site-packages instead of lib/python to match convention.
By setting the --root and --install-lib options to setup.py install to
${CMAKE_INSTALL_PREFIX} and ${C_PYTHONARCHLIB}, respectively, the package
files end up in the correct locations corresponding to the so-files.

For tarball builds, this means users can pick up and use the bindings by
setting LD_LIBRARY_PATH=$PREFIX PYTHONPATH=$PREFIX/pythonX.Y/site-packages.

Now that proof of concept is complete for tarball builds of an old branch
(pre-building for both Python 2 and 3), this branch is pretty much dead
(as is htcondor_jobs as it was at this point in time), so work on this
ticket will continue in a new branch to be named later.