Better packaging of Python bindings #6907
1. Non-C++ code moved or added to /bindings/python, and this
directory can contain native Python code to extend the bindings.

2. Python's setuptools are now used to install the bindings'
scaffolding, including an egg-info directory, so that pip
recognizes the bindings' install and version statuses.

3. On Linux and MacOS, the "suffix" of the SO files are now named
based on the output of Python's "sysconifg.get_config_var('SO')",
matching what we already do in Windows. This provides unique names
to SO files built for use with Python 3, which means no more renaming
of files during the build step.

4. Wheel builds are more simplified, and now when
"-DWANT_PYTHON_WHEELS:BOOL=ON" is passed to cmake, a wheel is
built automatically under CONDOR_SRC/bindings/python/dist.

5. Attempted to add all the changes above to the SRPM spec file,
and changed hardcoded Python 3 paths in spec file to use
%{python3_other_sitelib} macro provided by python3-rpm-macros
package (and which points to the python3.6 lib dir).

6. Attempted to add all the changes above to the new-debian
build rules. Probably need some help to get this correct for
release.

7. Cleaned up setup.py (which is now a cmake template) and
updated the README to point to htcondor.readthedocs.io.