Bugfix: Have both sides use same encryption setting. #7289
When the last file transferred is encrypted - but encryption is otherwise
disabled for a file transfer - then the upload side would send the final
transfer ACK encrypted while the download side would expect the final
ACK to be unencrypted.  This mismatch would subsequently cause the entire
file transfer to fail.  This bugfix simply sets the encryption back to
the default at the same time.

It's unclear when this was introduced due to lack of git history predating
2006; the bug is likely at least 13 years old.