Fixup F27 warning. #7318
It's not entirely clear whether the std::move would return a
local reference; this replaces the usage of rvalue refs with RVO.

Also fix buffer sizing issues; it's not clear if these were actually
able to trigger given the loop constraints.  Will trust the compiler's
advice on that one, however.