Utilize a manifest file for data reuse. #7389
This allows a user to create a file (the format is the same as the
output of the `sha256sum` binary) containing a list of files, one
per line, and the corresponding checksum.

HTCondor should transfer each file in the manifest as part of the
job and attempt to utilize the reuse mechanism for each.