(#7384) templatize ranger on element type to support job-ids
The old ranger is now `ranger<int>`, and we now specifically support
`ranger<JOB_ID_KEY>`.


In general, any TYPE is possible, with a few small requirements.

Currently the following expressions must be valid:

        TYPE a;
        TYPE b = a;
        int n;

        a < b;
        a == b;
        a + n;
        a - n;
        ++a;
        --a;

Which, for a custom TYPE, means implementing the following:

  - bool operator<(...)   // member function or global
  - bool operator==(...)  // member function or global
  - TYPE operator+(int)
  - TYPE operator-(int)
  - TYPE &operator++()
  - TYPE &operator--()
  - TYPE()                // default constructor to "zero" state
  - TYPE(const TYPE &x)   // copy constructor


Since we only require the `==` and `<` operators from the type, we have
to be careful in the header to rewrite any comparisons in terms of
these.  As for the .cpp, we just use the std::rel_ops namespace to avoid
contorting otherwise more natural-looking expressions.


Note that since this is now a template class, there is the question
about where the template implementation lives.  The most generic way is
to put it all into the header file, but I am not real wild about all of
that, plus the supporting static functions, being brought into a header.

So instead, I've opted to leave the implementation in the .cpp file,
with one explicit template instantiation per element type desired to be
used.  Currently that is:

        template struct ranger<int>;
        template struct ranger<JOB_ID_KEY>;

Which we'll need to add for every type we might want to use.


Additionally, for every TYPE, the following static methods need to be
defined for use by the load/persist methods:

        // extract element from string s, output to out,
        // return pos after element
        char *read_element(const char *s, TYPE *out);

        // write 'in' to buf, return number of bytes written
        int write_element(TYPE in, char *buf);


Relatedly, the load/persist functions have been moved into the ranger
class proper, as they needed to be templatized on the same type
parameter as the ranger class anyway.


...


New tests are added for ranger<JOB_ID_KEY>.  As the number of groups of
tests has grown, I've factored out more of the common bits, both to cut
down on redundant lines, and (maybe more importantly) to avoid mistakes
in setting up all the test table logic each time.