Remove executable checksum. #7389
This removes the ability to set a hash for data reuse of the executable.
That functionality was meant to simply be scaffolding for testing of
the data reuse directory; now that there is the data reuse manifest,
it is no longer needed.

The executable checksum functionality was never in a release; this
removal needs no release notes.