Bugfix: invoke stageout plugin at job end.  #7078
When `OutputPluginsOnlyOnExit` is set to `true` (the default value; there
would have been no reason why someone would set the undocumented knob
to this value either), output plugins were not being invoked on exit.
This corrects the logic to ensure plugins are invoked on exit for all
cases.

This additionally refactors the stageout logic as suggested in #7042
to make it simpler to read.