Fix negotiator not matching IPv6 startds to dual-stack schedds. #7397
The getSinfulStringProtocolBools() call for the schedd sinful was
short-circuiting after finding the preferred protocol (usually IPv4).
Thus, startds that only support the non-preferred protocol would never
be matched.
Fix the call to not short-circuit for the schedd and add a comment about
the behavior of the relevant arguments to
getSinfulStringProtocolBools().