Initial code for nonblocking TCP updates to the view collector. #6249
When in nonblocking mode, if the TCP socket to the view collector fills
up, it is registered with DaemonCore and subsequent updates are queued
locally. When the backlog clears, the queue updates are then sent.
There is a cap on the number of queued updates, after which additional
ones are dropped. The config params UPDATE_VIEW_COLLECTOR_NONBLOCKING
and UPDATE_VIEW_COLLECTOR_MAX_BACKLOG control this behavior.

Planned improvements:
* Where possible, store references to daemon ads for the queued updates
instead of copies of the udpate ads.
* When multiple updates to the same daemon ad are queued, discard all
but the latest (or merge the updates if diff ads become a thing).