Add SUFFIX as a build arg
This is most important so that the execute, cm, and submit images, which
are derived from the base images, use the correct tag for the base
image.  That is, if the base is tagged htcondor/base-el7-test, then
the execute image must do FROM htcondor/base-el7-test in order to get
the right contents.