Fix flocking when negotiator prefetch is on #7549
CHTC reports that the number of flocking out jobs has almost completely
gone to zero. Investigation shows that with negotiator rrl prefetch on,
(now the default), the negotiator sends an END NEGOTIATE command after
the prefetch, as well as after the negotiation proper. When the schedd
gets each one, it recalculates the flock level. After an END NEGOTIATE
after prefetch, no jobs have been matched, and none have been sent. The
schedd treats this as if the user has been completely satisfied, and
then lowers the flock level. As a result, the flock level is frequently
knocked back to zero, and no jobs flock out.

To fix this, we will notice when the end of negotiation message really
 means end of rll fetch, and ignore the rest of the end of negotiation
processing. This should also remove some load from the schedd.