fix for #7591: preempting claim in a GPU slot attempts to add another GPU and fails
when preempting, the resources are deleted from the d-slot and then re-added (sigh) for the new job
This hack is incomplete though, when the GPU count is decremented, the assigned GPU is not released,
so a later attempt to bind a GPU into a slot will fail because the count of free GPUs no longer
matches the number of items in the list of free GPUS.  The fix is to unbind the GPU at the same
time that we decrement the GPU count in the d-slot.  Note that this means that the next job can
end up being assigned a different GPU, but since GPUs are currently required to homogeneous, this is ok
for now.

Signed-off-by: John (TJ) Knoeller <johnkn@cs.wisc.edu>