Fix policy in security sessions received by negotiator from old schedds #7615
When an 8.9.3 schedd creates a non-negotiated security session for use
by the negotiator to contact it, the policy it sends out doesn't include
the encryption algorithm, and the default algorithm changed in 8.9.4
from 3DES to BLOWFISH. This causes breakage when a newer negotiator
tries to talk to an 8.9.3 schedd.
Now, the negotiator inserts the old default algorithm into the policy
(CryptoMethods="3DES") when importing the security session from an 8.9.3
schedd.