Fix mismatch of gahp ids when using remote_gahp #7623
The INFNBatchJob and INFNBatchResource objects used different IDs for
the batch gahp child process when the remote_gahp was used with
additional arguments (e.g. when --rgahp-key is used to specify a
different ssh key to use). This caused the Resource object to think the
batch_gahp never started.
The only practical effect was an annoying log message printed
repeatedly.