Keep a string version of the request ID. #7641
If we convert to an integer, we strip off the leading zeros; keep
a copy of the string version so, when a user requests this request
ID later on, we can do an exact string match instead of mucking around
with integers.

Fixes a bug where `condor_token_request_list` wouldn't filter correctly
on requests with a leading `0`.