Defer reporting errors until commit time. #7662
Currently, the behavior in NoAck mode is to immediately close the
TCP socket when errors occur during SetAttribute - providing no
opportunity to pass along an error message from the failed operation.

With this, the error message is buffered until an attempted commit
is made.  At that point, the transaction is aborted.  Once an attribute
is failed, all subsequent NoAck operations are ignored in that
transaction.