Do not refuse submit for UID / GID 0. #7662
Increasingly, the submitting user in the schedd is decoupled from
the Unix UID / GID that invoked condor_submit.  This is especially
true with remote submission cases and non-FS-authentication.

Now that there are reasonable error messages coming back from the
schedd when the job's Owner attribute is set to root (which is the
real thing we want to prevent), we can remove this check.