Clarify / refactor semantics of `User` and `Owner`.  #7648
This is the first step in refactoring the semantics of the `Owner`
and `User` attributes in preparation for cases where `Owner` does
not match the `User`.

In short, `User`:
- Becomes an immutable attribute "owner" by the schedd.
- Is set to the authenticate identity of the socket (unless overridden
  by a super-user).
- Is used for the various permissions check -- except for superusers,
  the socket's authenticated identity must match the `User` attribute
  for the purpose of modifying jobs.
- Remains the default value for the identity considered by the accountant
  and matchmaker.

On the other hand, `Owner`:
- Is the Windows / Unix username for launched processes and owners
- Remains the target of the various QUEUE_SUPER_USER_MAY_IMPERSONATE
  rules (as this protects the OS resources from a super user).

This commit has been tested for basic functionality (submit, hold,
release, negotiation, running jobs to completion) but hasn't endured
a full battery of tests.  There's some new code added for having
`User` be in a different UID domain but this is yet untested.  The
goal here is to change the semantics and then spend some time looking
for places where the abstraction leaks.  Once that's done, the new
features will be tested / improved.