Add authentication methods for non-negotiated security sessions. #7683
Define special authentication methods for non-negotiated security
sessions. These become the AuthMethods attribute of the security policy,
and the AuthenticationMethod attribute when advertising to the
collector. Since none of the regular authentication methods are
relevant, we define new names that reflect the relationship. Currently,
we have FAMILY and MATCH.
Previously, the value was the list of configured authentication methods
to try when negotiating a regular security session.