Fix usage of effective FQU. #7648
The "effective" FQU was not used consistently throughout the QMGMT
code.  This helps fix cases where a superuser is impersonating and
making changes on others' behalf.

Particularly, this also fixes the usage for superusers that authenticate
with the various condor@{family,parent,child} identities; it assumes
these superusers should be treated as being in the same UID domain.