Make flakey test more reliable. #7648
This test relied on jobs leaving the queue immediately when a
`condor_rm` is executed.  Periodically, I'd observe target jobs
actually getting started by the shadow - meaning it took a few
hundred milliseconds to preempt them and get them to leave the
queue.

By submitting the job on hold, a shadow is never started and the
race condition is much harder to trigger.