Fix negotiator priority level in collector with family session. #7729
Before, when using the non-negotiated family security session, the
negotiator was given low priority in the collector (because it thought
the client subsystem was collector).
Now, the collector gives high prority to queries where it thinks the
client subsystem is collector (this primarily means anyone using the
family security session).
Yes, this is a hack.