Fix GetReferences() for subscript notation. #7750
When an external reference refers to an attribute in a nested ad that's
inside a list (e.g. B[3].C), GetReferences() returns 'B[3]' instead of
'B'.
ClassAd::GetExternalReferences() returns the full attribute reference,
as asked for. TrimReferenceNames() in condor_utils needs to look for '['
in addition to '.' when trimming the name down.