Change mapping inputs to be the accounting domain #7793
After some discussions, it became apparent that the security subsystem
was the wrong place to hang the mapping on.  In reality, we want to
firm up the concept of an *accounting domain*.  That is, the namespace
of submitters to a particular negotiator.  Given the submitters are the
same at the schedd side for all collectors it flocks to, we want the
negotiator to be able to map submitters in a non-local accounting to
a group in the local domain (the map here is not general purpose - still
just adding a prefix).

This commit introduces the concept of ACCOUNTING_DOMAIN (defaults to
UID_DOMAIN), is tested with both old-and-new logic for the `User` attribute,
and moves the mapping mechanism to it.